Добавлена адаптация баннера, вынесен в отдельный компонент, убраны лишние div'ы в page.js
