+++ sultan/app/page.css
@@ -1,164 +1,20 @@
 .container {
-    width: 67%;
-    max-width: 1367px;
-    margin: 0 auto;
-    @media screen and (max-device-width: 768px) {
-        width: 90%;
-        padding-left: 15px;
-        padding-right: 15px;
-      }
-}
-.section { /*Разграничил секции, при их реализации стоит убрать отступы и границы */
-    padding: 4px;
-    margin: 4px;
-    z-index: 1;
-    @media screen and (max-device-width: 768px) {
-        margin-top: 0;}
-}
-.banner {
-    height: 796px;
-    width: 100%;
-    position: relative;
-    overflow: visible;
-}
-
-.banner_image {
-    background-image: url('../public/banner_home.webp');
-    position: absolute;
-    top: 0;
-    left: -50%;
-    width: 200%; 
-    height: 100%;
-    background-size: cover;
-    background-position: center;
-    z-index: -1;
-}
-.blur {
-    height: inherit;
-    backdrop-filter: blur(13px); /* Эффект размытия фона */
-    mask-image: linear-gradient(to right, black, transparent);
-    background: linear-gradient(90deg, rgba(138, 148, 171, 1) 0%, rgba(138, 148, 171, 0) 100%);
-    opacity: 1;
-}
-.banner_info {
-    width: 67%;
-    max-height: 796px;
-    max-width: 1367px;
-    
-    color: #fff;
-    white-space: nowrap;
-
-    margin: 0 auto;
-    margin-left: 8px;
-    margin-top: 98px;
-    margin-bottom: 94px;
-}
-
-.banner_info h1 {
-
-    font-size: 70px;
-    font-weight: 500;
-    line-height: 110%;
-}
-.banner_info h2 {
-    margin-top: 40px;
-    margin-left: 5px;
-
-    font-size: 30px;
-    font-weight: 700;
-    line-height: 150%;
-    letter-spacing: 0;
-    text-transform: uppercase;
-}
-.banner_info h2 span {
-    font-weight: 800;
-}
-
-.banner_info .btn {
-    width: 300px;
-    height: 95px;
-    margin-top: 60px;
-    margin-left: 5px;
-
-    background-color: #FFC85E;
-    border-radius: 80px;
-    border: none;
-    transition: background-color 0.3s ease;
-
-    color: #fff;
-    font-size: 18px;
-    font-weight: 700;
-    letter-spacing: 4px;
-
-
-}
-
-/* Стили при наведении курсора */
-.banner_info .btn:hover {
-    background-color: #ffa000;
-    cursor: pointer;
-}
-
-/* Стили при нажатии */
-.banner_info .btn:active {
-    background-color: #FFA500; /* Цвет фона при нажатии */
-    transform: scale(0.99); /* Уменьшаем размер кнопки при нажатии */
-}
-
-
-.banner_info .list {
-    display: flex;
-    justify-content: space-between;
-    flex-wrap: wrap;
-
-    width: 540px;
-    height: auto;
-
-    margin-top: 90px;
-}
-
-.banner_info .advant {
-    width: 250px;
-    display: flex;
-    flex-wrap: nowrap;
-    
-}
-.banner_info .circle {
-    width: 39px; /* Ширина кружка */
-    height: 39px; /* Высота кружка */
-    background-color: #FFC85E;
-    border-radius: 50%; /* Округление до круга */
-    display: flex; /* Используйте flex для центрирования содержимого */
-    align-items: center; /* Центрируем по вертикали */
-    justify-content: center; /* Центрируем по горизонтали */
-    margin: auto ;
-    margin-right: 10px;
-}
-.circle b {
-    font-size: 20px;
-    cursor: default;
-    user-select: none;
-}
-
-.banner_info .advant span {
-    font-size: 14;
-    font-weight: 400;
-    line-height: 150%;
-}
-
-.mini-card-container {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 20px;
+  width: 67%;
+  max-width: 1367px;
+  margin: 0 auto;
+  @media screen and (max-device-width: 768px) {
+    width: 90%;
+    padding-left: 15px;
+    padding-right: 15px;
   }
-  
-  .mini-card {
-    cursor: pointer;
-    transition: transform 0.3s ease;
+}
+.section {
+  padding: 4px;
+  margin: 4px;
+  z-index: 1;
+  @media screen and (max-device-width: 768px) {
+    margin-top: 0;
+    padding: 0px;
+    margin: 0px;
   }
-  
-  .mini-card:hover {
-    transform: scale(1.05);
-  }
-
-
+}
