+++ sultan/app/shared/banner/banner.css
@@ -1,0 +1,262 @@
+.banner { 
+  position: relative;
+  width: 100%;
+  height: 796px;
+  overflow: visible;
+
+  .banner_image {
+    position: absolute;
+    top: 0;
+    left: -50%;
+    width: 200%;
+    height: 100%;
+    z-index: -1;
+    overflow: hidden;
+
+    .banner_img {
+      position: absolute;
+      top: 0;
+      left: 0;
+      width: 100%;
+      height: 100%;
+      object-fit: cover;
+      z-index: -2;
+    }
+
+    .blur {
+      position: absolute;
+      top: 0;
+      left: 0;
+      width: 100%;
+      height: 100%;
+      backdrop-filter: blur(13px);
+      mask-image: linear-gradient(to right, black, transparent);
+      background: linear-gradient(90deg, rgba(138, 148, 171, 1) 0%, rgba(138, 148, 171, 0) 100%);
+      opacity: 1;
+    }
+  }
+
+  .banner_info {
+    position: relative;
+    z-index: 1;
+    width: 67%;
+    max-width: 1367px;
+    max-height: 796px;
+    color: #fff;
+    margin-left: 8px;
+    padding-top: 98px;
+    white-space: nowrap;
+
+    .mobile {
+      display: none;
+    }
+
+    h1 {
+      font-size: 70px;
+      font-weight: 500;
+      line-height: 110%;
+    }
+
+    h2 {
+      margin-top: 40px;
+      margin-left: 5px;
+      font-size: 30px;
+      font-weight: 700;
+      line-height: 150%;
+      text-transform: uppercase;
+
+      span {
+        font-weight: 800;
+      }
+    }
+
+    .btn {
+      width: 300px;
+      height: 95px;
+      margin-top: 60px;
+      margin-left: 5px;
+      background-color: #ffc85e;
+      border-radius: 80px;
+      border: none;
+      color: #fff;
+      font-size: 18px;
+      font-weight: 700;
+      letter-spacing: 4px;
+      transition: background-color 0.3s ease;
+
+      &:hover {
+        background-color: #ffa000;
+        cursor: pointer;
+      }
+
+      &:active {
+        background-color: #ffa500;
+        transform: scale(0.99);
+      }
+    }
+  }
+
+  @media screen and (max-width: 768px) {
+    height: auto;
+
+    .banner_image {
+      position: absolute;
+      left: -6vw;
+      width: 102vw;
+      height: auto;
+      aspect-ratio: 31 / 21;
+      overflow: visible;
+
+      .banner_img {
+        object-position: center;
+      }
+
+      .blur {
+        display: none;
+      }
+    }
+
+    .banner_info {
+      padding-top: 3vw;
+      margin-left: 0;
+      width: 100%;
+      height: auto;
+      aspect-ratio: 31 / 21;
+      display: flex;
+      flex-direction: column;
+      justify-content: center;
+
+      .mobile {
+        display: inherit;
+      }
+
+      .desktop {
+        display: none;
+      }
+
+      .mblur {
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+        align-items: center;
+        position: relative;
+        text-align: center;
+        background: linear-gradient(90deg, rgba(138, 148, 171, 1) 0%, rgba(138, 148, 171, 0) 100%);
+        aspect-ratio: 29 / 10;
+        width: 100%;
+        backdrop-filter: blur(4px);
+      }
+
+      h1 {
+        font-size: 18px;
+        font-weight: 500;
+        line-height: 110%;
+      }
+
+      h2 {
+        margin-top: 0;
+        margin-left: 0;
+        font-size: 10px;
+        font-weight: 500;
+        line-height: 150%;
+        text-transform: uppercase;
+
+        span {
+          font-weight: 500;
+        }
+      }
+
+      .btn {
+        display: none;
+      }
+    }
+  }
+
+  .list {
+    display: flex;
+    justify-content: space-between;
+    flex-wrap: wrap;
+    width: 540px;
+    margin-top: 90px;
+
+    &.desktop-pluses {
+      display: flex;
+    }
+
+    &.mobile-pluses {
+      display: none;
+    }
+
+    .advant {
+      display: flex;
+      flex-wrap: nowrap;
+      width: 250px;
+
+      .circle {
+        width: 40px;
+        height: 40px;
+        background-color: #ffc85e;
+        border-radius: 50%;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        margin: auto 10px;
+        
+        b {
+          font-size: 20px;
+          cursor: default;
+          user-select: none;
+        }
+      }
+
+      span {
+        font-size: 14px;
+        font-weight: 400;
+        line-height: 150%;
+      }
+    }
+
+    @media screen and (max-width: 768px) {
+      justify-content: flex-start;
+      flex-direction: column;
+      flex-wrap: nowrap;
+      width: 100%;
+      margin-top: 8vw;
+
+      &.desktop-pluses {
+        display: none;
+      }
+
+      &.mobile-pluses {
+        display: flex;
+      }
+
+      .advant {
+        width: 100%;
+        margin-top: 10px;
+
+        .circle {
+          width: 20px;
+          height: 20px;
+          background-color: #ffc85e;
+          border-radius: 50%;
+          display: flex;
+          align-items: center;
+          justify-content: center;
+          margin-right: 5px;
+
+          b {
+            font-size: 10px;
+            color: #fff;
+          }
+        }
+      }
+
+      span {
+        font-size: 10px;
+        color: #3F4E65;
+        letter-spacing: 0;
+      }
+    }
+  }
+}
